mobil_backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
httpx
python-multipart
